Struct kickstarter::db::kickstarter::project::Project

pub struct Project {
    pub project_id: i32,
    pub name: String,
    pub date_created: NaiveDateTime,
    pub goal: f64,
    pub pledge: Vec<Pledge>,
}

Fields

project_id

primary default: nextval('kickstarter.project_project_id_seq'::regclass) not nullable db data type: integer

name

unique not nullable db data type: text

date_created

default: ('now'::text)::timestamp without time zone not nullable db data type: timestamp without time zone

goal

not nullable db data type: double precision

pledge

has many

Methods

impl Project

fn create(client: &Client, project_name: &str, amount: f64) -> Result<Project>

Creates a new Kickstarter project with the provided goal amount in dollars. Returns the created project on success.

fn get_id(client: &Client, project_name: &str) -> Result<Value>

Retrieve a project ID by name.

fn list_all(client: &Client) -> Result<Vec<Project>>

Returns a list of all projects on Kickstarter.

fn list_backers(client: &Client, project_name: &str) -> Result<(BTreeMap<User, f64>, f64)>

Retrieves a list of all users that have backed a given project. Returns a map of User objects to their contributions, as well as the overall project goal amount.