Struct kickstarter::db::kickstarter::user::User

pub struct User {
    pub user_id: i32,
    pub name: String,
    pub date_created: NaiveDateTime,
    pub pledge: Vec<Pledge>,
}

Fields

user_id

primary default: nextval('kickstarter.user_user_id_seq'::regclass) not nullable db data type: integer

name

unique not nullable db data type: text

date_created

default: ('now'::text)::timestamp without time zone not nullable db data type: timestamp without time zone

pledge

has many

Methods

impl User

fn upsert(client: &Client, user: &str) -> Result<Value>

Upserts a user and returns the resultant ID as a Value.

fn list_pledges(client: &Client, user: &str) -> Result<BTreeMap<Project, Pledge>>

Retrieve a map of all pledges that a user has made to Kickstarter projects. Returns a map of project names to Pledge objects.