Module kickstarter::validate

Kickstarter data validations.

Reexports

pub use self::error::{Error, Result};

Modules

error
Validation error module

Functions

alphanumeric
Validates that the string is alphanumeric and contains only underscores and dahes as special characters.

currency
Validates that the float is positive, and rounds it to two decimal places.

length
Validates that the string length is between min and max, inclusive.

luhn10
Validates that a numerical string passes the Luhn-10 test.

numtext
Validates that the string contains only digits.

regex
Validates that the string matches the provided regex.